refactor: Move TableNode type definitions to a separate file and clean up imports
