+++ frontend/packages/erd-core/src/components/ERDRenderer/ERDContent/TableNode/index.ts
@@ -1,1 +1,2 @@
 export * from './TableNode'
+export * from './type'
