+++ frontend/packages/erd-core/src/components/ERDRenderer/ERDContent/TableNode/type.ts
@@ -1,0 +1,12 @@
+import type { Table } from '@liam-hq/db-structure'
+import type { Node } from '@xyflow/react'
+
+export type Data = {
+  table: Table
+  isHighlighted: boolean
+  isRelated: boolean
+  highlightedHandles: string[]
+  sourceColumnName: string | undefined
+}
+
+export type TableNodeType = Node<Data, 'table'>
